+++ src/Storage/Query/Directive/HydrateDirective.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Bolt\Storage\Query\Handler;
+namespace Bolt\Storage\Query\Directive;
 
 use Bolt\Storage\Query\QueryInterface;
 
 /**
- *  Handler to add a limit modifier to the query.
+ *  Directive to add a limit modifier to the query.
  */
-class HydrateHandler
+class HydrateDirective
 {
     /**
      * @param QueryInterface $query
